+++ xasm/qasm/expression.h
@@ -14,17 +14,17 @@ typedef struct Expression {
 	bool isConstant;
     union {
         long double floating;
-        int32_t integer;
+        int64_t integer;
         struct Symbol* symbol;
     } value;
 } SExpression;
 
 
 extern SExpression*
-expr_Const(int32_t value);
+expr_Const(int64_t value);
 
 extern SExpression*
-expr_CheckRange(SExpression* expression, int32_t low, int32_t high);
+expr_CheckRange(SExpression* expression, int64_t low, int64_t high);
 
 extern void
 sect_OutputExpr8(struct Expression* expr);
